Wrap savePolicy in a transaction
Inlined `dropTable` function into the body of `savePolicy` to make
wrapping in a transaction cleaner - it was not used elsewhere, and was
misleadingly named (it in fact _truncates_ the casbin table).

Removed call to `createdTable` in the `savePolicy` function. As the
table is not dropped there is no need to call `sync` again at this
point.